chore(ci): further pre-commit cleanup
